1.更新部分代码

+++ lib/DevApp/src/main/java/dev/utils/common/ArrayUtils.java
@@ -1760,7 +1760,7 @@ public final class ArrayUtils {
                 JCLogUtils.eTag(TAG, e, "get");
             }
         }
-        return (long) -1;
+        return -1L;
     }
 
     /**
